Bug 624842: update build scripts to handle *.abc being located in 'generated' instead of 'core/builtin.abc' and 'shell/shell_toplevel.abc' - additional tweaks to other build files to handle the new location of the generated abc/cpp/h files (r+stejohns)

+++ core/builtin.py
@@ -48,7 +48,8 @@ def mv(oldfile, newfile):
     os.remove(oldfile)
 
 def rm(file):
-    os.remove(file)
+    if os.access(file, os.F_OK) == True:
+        os.remove(file)
 
 classpath = os.environ.get('ASC')
 if classpath == None:
